Simplify and refactor OpenStreetMap component configuration
