+++ src/wwElement.vue
@@ -1,11 +1,5 @@
 <template>
-  <div v-if="error" class="map-placeholder">
-    <div class="placeholder-content">
-      {{ error }}
-    </div>
-  </div>
   <div
-    v-else
     class="ww-leaflet leaflet-container leaflet-touch leaflet-retina leaflet-fade-anim leaflet-grab leaflet-touch-drag leaflet-touch-zoom"
     :class="{ editing: isEditing }"
     ref="mapContainer"
@@ -14,9 +8,9 @@
 </template>
 
 <script>
-import { ref, onMounted, onBeforeUnmount, nextTick, watch } from "vue";
-import { useLeafletMap } from "./useLeafletMap";
+import { ref, computed, onMounted, onUnmounted, watch, nextTick } from "vue";
 import "leaflet/dist/leaflet.css";
+import useLeafletMap from "./use/useLeafletMap";
 
 export default {
   name: "OpenStreetMap",
@@ -29,10 +23,6 @@ export default {
   emits: ["trigger-event"],
   setup(props, { emit }) {
     const componentKey = ref(0);
-    const error = ref("");
-    const mapContainer = ref(null);
-    let mapInstance = null;
-
     /* wwEditor:start */
     const isMarkersBound = computed(() => {
       return !!props.wwEditorState.boundProps.markers;
@@ -74,6 +64,9 @@ export default {
     };
     /* wwEditor:end */
 
+    let mapInstance = null;
+    const mapContainer = ref(true);
+
     function debounce(func, wait) {
       let timeout;
       return function (...args) {
@@ -87,34 +80,17 @@ export default {
     }
 
     async function initMap() {
-      try {
-        error.value = "";
-        componentKey.value += 1;
-        await nextTick();
+      componentKey.value += 1;
+      await nextTick();
 
-        const { map } = useLeafletMap(
-          mapContainer.value,
-          props.content,
-          boundStates,
-          emit
-        );
+      const { map } = useLeafletMap(
+        mapContainer.value,
+        props.content,
+        boundStates
+        emit
+      );
 
-        mapInstance = map;
-
-        // Initialize map controls based on props
-        if (props.content.scaleControl) {
-          L.control.scale().addTo(mapInstance);
-        }
-
-        if (props.content.layerControl) {
-          const baseLayers = {};
-          const overlays = {};
-          L.control.layers(baseLayers, overlays).addTo(mapInstance);
-        }
-      } catch (error) {
-        console.error("Map initialization error:", error);
-        error.value = error.message;
-      }
+      mapInstance = map;
     }
 
     const debouncedInitMap = debounce(initMap, 1000);
@@ -166,45 +142,12 @@ export default {
       { deep: true }
     );
 
-    return {
-      isEditing,
-      mapContainer,
-      mapInstance,
-      componentKey,
-      error,
-    };
+    return { isEditing, mapContainer, mapInstance, componentKey };
   },
 };
 </script>
 
 <style lang="scss" scoped>
-.map-placeholder {
-  overflow: hidden;
-  z-index: 2;
-  position: absolute;
-  top: 0px;
-  left: 0px;
-  height: 100%;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  background-color: rgba(0, 0, 0, 0.4);
-
-  .placeholder-content {
-    text-align: center;
-    width: 90%;
-    background: white;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    padding: 0.8em 1.2em;
-    border-radius: 12px;
-  }
-}
-
 .ww-leaflet {
   width: 100%;
   overflow: hidden;
